test: satisfy corsa bind update gates
