+++ davinci-road/plan/assertion-allowlist.toml
@@ -200,7 +200,6 @@ paths = [
   "crates/vize_canon/tests/tier_l_incremental.rs",
   "crates/vize_canon/tests/tsconfig_option_diagnostics.rs",
   "crates/vize_carton/src/config/model/vue.rs",
-  "crates/vize_carton/src/corsa_resolver.rs",
   "crates/vize_carton/src/flags.rs",
   "crates/vize_carton/src/source_range.rs",
   "crates/vize_carton/tests/loading.rs",
@@ -244,7 +243,6 @@ paths = [
   "crates/vize_marquette/tests/cli.rs",
   "crates/vize_marquette/tests/test_run_schema.rs",
   "crates/vize_patina/src/linter/corsa_session/paths.rs",
-  "crates/vize_patina/src/linter/corsa_session/session.rs",
   "crates/vize_patina/src/linter/native_type_aware/source_path.rs",
   "crates/vize_patina/src/linter/script_rules/html_scripts.rs",
   "crates/vize_patina/src/markup/tests.rs",
